requirements.txt
# Core
watchdog>=2.1.6
click>=8.1.3
python-dotenv>=0.19.0

# OCR & Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
imagehash>=4.3.1

# System & Context
psutil>=5.9.0
pygetwindow>=0.0.9

# NLP & ML
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2

# Database & Search
sqlalchemy>=2.0.0
whoosh>=2.7.4

# Cloud Storage
google-api-python-client>=2.80.0
google-auth-oauthlib>=1.0.0
dropbox>=11.36.0

# UI
streamlit>=1.25.0
Pillow>=10.0.0
